example/transformers/models/rag/requirements.txt
gradio==4.32.2
llama-index==0.10.43
llama-index-embeddings-huggingface==0.2.1
llama-index-vector-stores-chroma==0.1.8
llama-index-vector-stores-faiss==0.1.2
faiss-cpu==1.8.0
